Add files via upload
The People living in Agon Community in Obi Local Government Area of Nasarawa State in Nigeria, are constant in need of good water. The community which have over 15,000 people have suffered for a long time geting good water. Our heart desire is to raise partners who can join hands with us and provide good water to ruler Communities who lack one. We want to drill a well with over head tank. The Project is to provide over 15,000 people with good water, The provide will need 1,500,000 naira. We are just about to start this project but we need to raise enough partners before we can start. If you are move to partner we us to provide good water you can email us on: johnbisong4444@gmail.com or call +2347062074421. Remember is healthy to drink save water.